@@ -19,7 +19,7 @@ public struct ASRConfig: Sendable {
 
 // MARK: - Results
 
-public struct ASRResult: Sendable {
+public struct ASRResult: Codable, Sendable {
     public let text: String
     public let confidence: Float
     public let duration: TimeInterval
@@ -46,7 +46,7 @@ public struct ASRResult: Sendable {
     }
 }
 
-public struct TokenTiming: Sendable {
+public struct TokenTiming: Codable, Sendable {
     public let token: String
     public let tokenId: Int
     public let startTime: TimeInterval
