@@ -1,6 +1,6 @@
 import AVFoundation
 import Accelerate
-import CoreML
+@preconcurrency import CoreML
 import Foundation
 import OSLog
 
